+++ src/main/java/com/jdte/JDTE.java
@@ -100,6 +100,7 @@ public class JDTE {
         NeoForge.EVENT_BUS.addListener(ExtendedTimeAccelerationManager::onServerTickPost);
         NeoForge.EVENT_BUS.addListener(ExtendedTimeAccelerationManager::onLevelUnload);
         NeoForge.EVENT_BUS.addListener(ExtendedTimeAccelerationManager::onServerStopped);
+        NeoForge.EVENT_BUS.addListener(com.jdte.common.integrations.ae2.AE2CraftingReadNetwork::onServerStopped);
         NeoForge.EVENT_BUS.addListener(TimeAcceleratorIntegratedServerDefaults::onServerAboutToStart);
         NeoForge.EVENT_BUS.addListener(EventPriority.LOWEST, AEOutputManager::onServerTickPost);
         NeoForge.EVENT_BUS.addListener(EventPriority.LOWEST, MachineOutputManager::onServerTickPost);
@@ -230,6 +231,7 @@ public class JDTE {
     private void commonSetup(FMLCommonSetupEvent event) {
         event.enqueueWork(() -> {
             AEOutputNetwork.registerLinkable();
+            com.jdte.common.integrations.ae2.AE2CraftingReadNetwork.registerLinkable();
             com.jdte.common.integrations.curios.BigFluidTankCuriosIntegration.registerCurioBehaviors();
         });
     }