+++ src/main/java/com/jdte/common/upgrades/UpgradeType.java
@@ -13,6 +13,7 @@ public enum UpgradeType {
     FORTUNE("fortune", 8),
     PRECISION("precision", 1),
     AE_ACCELERATION("ae_acceleration", 1),
+    AE_CRAFTING_READ("ae_crafting_read", 1),
     AE_OUTPUT("ae_output", 1),
     ESSENCE_CONVERSION("essence_conversion", 1),
     SEED_CONVERSION("seed_conversion", 1);